Create a list of items Add a event listener Add items and save to the local storage
