+++ src/index.html
@@ -19,7 +19,7 @@
         </form>
         <ul class="list-container">
         </ul>
-        <p class="clear">Clear all completed</p>
+        <p class="clear"><span class="strikethrough">Clear all completed</span></p>
     </div>
 </body>
 </html>
